+++ toad/merge.pyx
@@ -105,7 +105,7 @@ def KMeansMerge(feature, target = None, nan = -1, n_bins = None, random_state = 
 
 
 
-def DTMerge(feature, target, nan = -1, n_bins = None, min_samples = 1):
+def DTMerge(feature, target, nan = -1, n_bins = None, min_samples = 1, **kwargs):
     """Merge by Decision Tree
 
     Args:
@@ -126,6 +126,7 @@ def DTMerge(feature, target, nan = -1, n_bins = None, min_samples = 1):
     tree = DecisionTreeClassifier(
         min_samples_leaf = min_samples,
         max_leaf_nodes = n_bins,
+        **kwargs,
     )
     tree.fit(feature.reshape((-1, 1)), target)
 
@@ -169,7 +170,7 @@ cpdef ChiMerge(feature, target, n_bins = None, min_samples = None,
     len_f = len(feature_unique)
     len_t = len(target_unique)
 
-    cdef double [:,:] grouped = np.zeros((len_f, len_t), dtype=np.float)
+    cdef double [:,:] grouped = np.zeros((len_f, len_t), dtype=float)
 
     for r in range(len_f):
         tmp = target[feature == feature_unique[r]]
@@ -194,7 +195,7 @@ cpdef ChiMerge(feature, target, n_bins = None, min_samples = None,
 
         # Calc chi square for each group
         l = len(grouped) - 1
-        chi_list = np.zeros(l, dtype=np.float)
+        chi_list = np.zeros(l, dtype=float)
         chi_min = np.inf
 
         for i in range(l):