+++ toad/metrics.py
@@ -4,7 +4,6 @@ from scipy.stats import ks_2samp
 
 from sklearn.metrics import f1_score, roc_auc_score, roc_curve
 
-
 from .utils import (
     feature_splits,
     iter_df,